Ratchet tree initial implementation

+++ src/util.rs
@@ -6,6 +6,6 @@ macro_rules! hex {
 }
 
 #[inline]
-pub fn log2(x: u64) -> u64 {
-    (x as f32).log2().floor() as u64
+pub fn log2(x: usize) -> usize {
+    (x as f32).log2().floor() as usize
 }
